+++ example/prometheus/main.go
@@ -23,8 +23,8 @@ import (
 	"time"
 
 	"go.opentelemetry.io/otel"
-	"go.opentelemetry.io/otel/api/global"
 	"go.opentelemetry.io/otel/exporters/metric/prometheus"
+	"go.opentelemetry.io/otel/global"
 	"go.opentelemetry.io/otel/label"
 )